+++ packages/x-scheduler-premium/src/internals/index.ts
@@ -1,10 +1,3 @@
-import { RecurrenceTab } from './components/event-dialog/RecurrenceTab';
-import { RecurringScopeDialog } from './components/scope-dialog/ScopeDialog';
-
-export { RecurrenceTab, RecurringScopeDialog };
-
-/** Slots that the premium scheduler injects into the community `EventDialogProvider`. */
-export const PREMIUM_EVENT_DIALOG_SLOTS = {
-  recurrenceTab: RecurrenceTab,
-  recurringScopeDialog: RecurringScopeDialog,
-};
+export { RecurrenceTab } from './components/event-dialog/RecurrenceTab';
+export { RecurringScopeDialog } from './components/scope-dialog/ScopeDialog';
+export { PREMIUM_EVENT_DIALOG_SLOTS } from './eventDialogSlots';
