+++ packages/x-scheduler/src/internals/components/index.ts
@@ -1,5 +1,6 @@
 export * from './event-drag-preview';
 export * from './event-dialog';
 
+export { EventCalendarProvider } from './EventCalendarProvider';
 export { EventCalendarRoot } from '../../event-calendar/EventCalendarRoot';
 export type { EventCalendarRootProps } from '../../event-calendar/EventCalendarRoot';
